requirements.txt
# Core
cryptography==41.0.7
pycryptodome==3.20.0

# Accumulator
gmpy2==2.1.5

# Async
aiofiles==23.2.1

# Storage
asyncpg==0.29.0
neo4j==5.17.0
redis==5.0.1
aiosqlite==0.19.0

# TEE
boto3==1.34.0

# Utils
pydantic==2.5.0
python-dateutil==2.8.2
